+++ src/Crowdin.Api/Dictionaries/DictionariesApiExecutor.cs
@@ -33,7 +33,7 @@ namespace Crowdin.Api.Dictionaries
         /// <a href="https://support.crowdin.com/enterprise/api/#operation/api.projects.dictionaries.getMany">Crowdin Enterprise API</a>
         /// </summary>
         [PublicAPI]
-        public async Task<ResponseList<Dictionary>> ListDictionaries(int projectId, string? languageIds = null)
+        public async Task<ResponseList<Dictionary>> ListDictionaries(long projectId, string? languageIds = null)
         {
             var url = $"/projects/{projectId}/dictionaries";
             
@@ -55,7 +55,7 @@ namespace Crowdin.Api.Dictionaries
         /// <a href="https://support.crowdin.com/enterprise/api/#operation/api.projects.dictionaries.patch">Crowdin Enterprise API</a>
         /// </summary>
         [PublicAPI]
-        public async Task<Dictionary> EditDictionary(int projectId, string languageId, IEnumerable<DictionaryPatch> patches)
+        public async Task<Dictionary> EditDictionary(long projectId, string languageId, IEnumerable<DictionaryPatch> patches)
         {
             var url = $"/projects/{projectId}/dictionaries/{languageId}";
             CrowdinApiResult result = await _apiClient.SendPatchRequest(url, patches);
